@@ -1,26 +1,23 @@
-apply plugin: 'com.android.application'
+plugins {
+    id 'com.android.application'
+}
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.2"
+    namespace "io.github.mthli.knifedemo"
+    compileSdk 35
 
     defaultConfig {
         applicationId "io.github.mthli.knifedemo"
-        minSdkVersion 14
-        targetSdkVersion 23
-        versionCode 2
-        versionName "1.1"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
+        minSdk 23
+        targetSdk 34
+        versionCode 6
+        versionName "2.0.0"
     }
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile project(':knife')
+    implementation "androidx.appcompat:appcompat:1.7.1"
+    implementation "androidx.coordinatorlayout:coordinatorlayout:1.3.0"
+    implementation "com.google.android.material:material:1.14.0-alpha09"
+    implementation project(":knife")
 }
